requirements.txt
appnope==0.1.0
bleach==2.1.2
decorator==4.2.1
Django==1.11.6
django-cors-headers==2.1.0
django-extensions==1.9.9
djangorestframework==3.7.3
djangorestframework-jwt==1.11.0
entrypoints==0.2.3
html5lib==1.0.1
ipykernel==4.8.1
ipython==6.2.1
ipython-genutils==0.2.0
ipywidgets==7.1.1
jedi==0.11.1
Jinja2==2.10
jsonschema==2.6.0
jupyter-client==5.2.2
jupyter-core==4.4.0
MarkupSafe==1.0
mistune==0.8.3
nbconvert==5.3.1
nbformat==4.4.0
notebook==5.4.0
pandocfilters==1.4.2
parso==0.1.1
pdfkit==0.6.1
pexpect==4.3.1
pickleshare==0.7.4
Pillow==5.0.0
prompt-toolkit==1.0.15
ptyprocess==0.5.2
Pygments==2.2.0
PyJWT==1.5.3
python-dateutil==2.6.1
pytz==2017.2
pyzmq==16.0.4
Send2Trash==1.4.2
simplegeneric==0.8.1
six==1.11.0
terminado==0.8.1
testpath==0.3.1
tornado==4.5.3
traitlets==4.3.2
typing==3.6.4
wcwidth==0.1.7
webencodings==0.5.1
widgetsnbextension==3.1.3
